+++ src/bam/record.jl
@@ -34,11 +34,14 @@ function Record(data::Vector{UInt8})
 end
 
 function Base.convert(::Type{Record}, data::Vector{UInt8})
+    length(data) < FIXED_FIELDS_BYTES && throw(ArgumentError("data too short"))
     record = Record()
-    unsafe_copy!(pointer_from_objref(record), pointer(data), FIXED_FIELDS_BYTES)
+    dst_pointer = Ptr{UInt8}(pointer_from_objref(record))
+    unsafe_copyto!(dst_pointer, pointer(data), FIXED_FIELDS_BYTES)
     dsize = data_size(record)
     resize!(record.data, dsize)
-    unsafe_copy!(pointer(record.data), pointer(data, FIXED_FIELDS_BYTES + 1), dsize)
+    length(data) < dsize + FIXED_FIELDS_BYTES && throw(ArgumentError("data too short"))
+    unsafe_copyto!(record.data, 1, data, FIXED_FIELDS_BYTES + 1, dsize)
     return record
 end
 
@@ -399,7 +402,9 @@ function cigar_position(record::Record, checkCG::Bool = true)::Tuple{Int, Int}
     if x != UInt32(seqlength(record) << 4 | 4)
         return cigaridx, nops
     end
-    tagidx = findauxtag(record.data, auxdata_position(record), UInt8('C'), UInt8('G'))
+    start = auxdata_position(record)
+    stop = data_size(record)
+    tagidx = findauxtag(record.data, start, stop, UInt8('C'), UInt8('G'))
     if tagidx == 0
         return cigaridx, nops
     end
@@ -563,12 +568,16 @@ end
 
 function Base.getindex(record::Record, tag::AbstractString)
     checkauxtag(tag)
-    return getauxvalue(record.data, auxdata_position(record), UInt8(tag[1]), UInt8(tag[2]))
+    start = auxdata_position(record)
+    stop = data_size(record)
+    return getauxvalue(record.data, start, stop, UInt8(tag[1]), UInt8(tag[2]))
 end
 
 function Base.haskey(record::Record, tag::AbstractString)
     checkauxtag(tag)
-    return findauxtag(record.data, auxdata_position(record), UInt8(tag[1]), UInt8(tag[2])) > 0
+    start = auxdata_position(record)
+    stop = data_size(record)
+    return findauxtag(record.data, start, stop, UInt8(tag[1]), UInt8(tag[2])) > 0
 end
 
 function Base.keys(record::Record)