+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Function convert(BAM.Record, x::Vector{UInt8} updated to Julia v. 1.0
 
 ## [1.0.0] - 2018-09-18
 ### Added